Fix BAZEL_RUST_STDLIB_LINKFLAGS
The expected behavior for BUILD_for_rust_toolchain() is for it to
receive a list of flags, and if None are specified, a default set of
link flags are selected dependent upon the target system.

The logic here was broken as any list specified would get silently
passed through, breaking further on in the code where it is expecting
stdlib_linkflags to be in string form.